requirements/test.txt
-r base.txt

dj-inmemorystorage == 1.4.1
factory_boy == 2.10.0
flake8 == 3.5.0
pytest == 3.4.0
pytest-django == 3.1.2
pytest-xdist == 1.22.0
